split cfgdlg part 3
